--- a/Manifest.toml
+++ b/Manifest.toml
@@ -1,0 +1,2 @@
+# This file is machine-generated - editing it directly is not advised
+