--- a/Manifest.toml
+++ b/Manifest.toml
@@ -1,2 +1,0 @@
-# This file is machine-generated - editing it directly is not advised
-