--- a/Project.toml
+++ b/Project.toml
@@ -1,4 +1,4 @@
 name = "ProceduralNoise"
 uuid = "114ee83c-a860-11e9-2720-319e06816c15"
-authors = ["NikulinVs"]
+authors = ["ararslan", "NikulinVs"]
 version = "0.1.0"