--- a/src/ProceduralNoise.jl
+++ b/src/ProceduralNoise.jl
@@ -4,5 +4,5 @@ include("interpolation.jl")
 include("perlin.jl")
 
 export smoothstep, smootherstep
-export perlin, octaveperlin
+export perlin, perlin3d, octaveperlin, octaveperlin3d
 end # module