Attempt to merge https://github.com/ararslan/ProceduralNoise.jl/pull/2.