@adrhill 2022-05-20
Due to type instability, the PR is currently slower by a factor of ~200. `perlin3d` is the old specialized code. 
```Julia
julia> @benchmark perlin($1.1, $2.2, $3.3)
BenchmarkTools.Trial: 10000 samples with 3 evaluations.
 Range (min … max):   8.805 μs …  3.553 ms  ┊ GC (min … max): 0.00% … 99.26%
 Time  (median):     10.980 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   12.312 μs ± 56.249 μs  ┊ GC (mean ± σ):  7.86% ±  1.72%

             ▄█   ▁                                            
  ▁▁▂▅▆▃▂▂▂▂▂██▆▃██▄▂▁▂▁▁▁▁▁▂▂▂▃▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▂
  8.81 μs         Histogram: frequency by time          19 μs <

 Memory estimate: 13.20 KiB, allocs estimate: 164.

julia> @benchmark perlin3d($1.1, $2.2, $3.3)
BenchmarkTools.Trial: 10000 samples with 982 evaluations.
 Range (min … max):  45.930 ns … 290.740 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     49.089 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   54.981 ns ±  17.501 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▆██▅▁    ▂▃▂▁    ▁▂▂▁                                        ▂
  ██████▇▇██████████████▇▇▆▇▇▇▆▇█▇▇▆▇▆▆▇▆▅▇▆▇▆▅▅▆▅▅▆▅▆▅▅▅▄▅▄▅▆ █
  45.9 ns       Histogram: log(frequency) by time       134 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```
Comparing the outputs, it looks like the generalized version has some artefacts:
```Julia
using ProceduralNoise
using ImageCore

rg = 0:0.01:20
A = Gray.([perlin(x, y) for x in rg, y in rg])
B = Gray.([perlin3d(x, y, 0.0) for x in rg, y in rg])
```

**New:**
![perlin_new](https://user-images.githubusercontent.com/20258504/169618041-2e5feba7-05c6-41e1-a433-5984d309b2fb.png)

**Old:**
![perlin_old](https://user-images.githubusercontent.com/20258504/169618049-a8a939b6-0dbc-487d-99d4-e9e618a652c6.png)